requirements.txt
Django === 6.0.1
psycopg2-binary==2.9.8
djangorestframework==3.16.1
django-filter==25.2
